♻️ refactor(vecs): consolidate vecs from quantity
