+++ src/coordinax/_src/vectors/d3/constructor.py
@@ -3,32 +3,30 @@
 
 __all__: list[str] = []
 
-from typing import Any
 
 from .base import AbstractAcc3D, AbstractPos3D, AbstractVel3D
-from .cartesian import CartesianAcc3D, CartesianPos3D, CartesianVel3D
 from coordinax._src.vectors.base import AbstractVector
 
 #####################################################################
 
 
-@AbstractVector.from_.dispatch(precedence=-1)
-def from_(cls: type[AbstractPos3D], obj: Any, /) -> CartesianPos3D:
-    """Try to construct a 3D Cartesian position from an object.
+# @AbstractVector.from_.dispatch(precedence=-1)
+# def from_(cls: type[AbstractPos3D], obj: Any, /) -> CartesianPos3D:
+#     """Try to construct a 3D Cartesian position from an object.
 
-    Examples
-    --------
-    >>> import unxt as u
-    >>> import coordinax as cx
+#     Examples
+#     --------
+#     >>> import unxt as u
+#     >>> import coordinax as cx
 
-    >>> x = u.Quantity([1, 2, 3], "km")
-    >>> vec = cx.vecs.AbstractPos3D.from_(x)
-    >>> print(vec)
-    <CartesianPos3D (x[km], y[km], z[km])
-        [1 2 3]>
+#     >>> x = u.Quantity([1, 2, 3], "km")
+#     >>> vec = cx.vecs.AbstractPos3D.from_(x)
+#     >>> print(vec)
+#     <CartesianPos3D (x[km], y[km], z[km])
+#         [1 2 3]>
 
-    """
-    return obj if isinstance(obj, CartesianPos3D) else CartesianPos3D.from_(obj)
+#     """
+#     return obj if isinstance(obj, CartesianPos3D) else CartesianPos3D.from_(obj)
 
 
 @AbstractVector.from_.dispatch(precedence=1)
@@ -58,23 +56,23 @@ def from_(cls: type[AbstractPos3D], obj: AbstractPos3D, /) -> AbstractPos3D:
 #####################################################################
 
 
-@AbstractVector.from_.dispatch(precedence=-1)
-def from_(cls: type[AbstractVel3D], obj: Any, /) -> CartesianVel3D:
-    """Try to construct a 3D Cartesian velocity from an object.
+# @AbstractVector.from_.dispatch(precedence=-1)
+# def from_(cls: type[AbstractVel3D], obj: Any, /) -> CartesianVel3D:
+#     """Try to construct a 3D Cartesian velocity from an object.
 
-    Examples
-    --------
-    >>> import unxt as u
-    >>> import coordinax as cx
+#     Examples
+#     --------
+#     >>> import unxt as u
+#     >>> import coordinax as cx
 
-    >>> x = u.Quantity([1, 2, 3], "km / s")
-    >>> vec = cx.vecs.AbstractVel3D.from_(x)
-    >>> print(vec)
-    <CartesianVel3D (d_x[km / s], d_y[km / s], d_z[km / s])
-        [1 2 3]>
+#     >>> x = u.Quantity([1, 2, 3], "km / s")
+#     >>> vec = cx.vecs.AbstractVel3D.from_(x)
+#     >>> print(vec)
+#     <CartesianVel3D (d_x[km / s], d_y[km / s], d_z[km / s])
+#         [1 2 3]>
 
-    """
-    return obj if isinstance(obj, CartesianVel3D) else CartesianVel3D.from_(obj)
+#     """
+#     return obj if isinstance(obj, CartesianVel3D) else CartesianVel3D.from_(obj)
 
 
 @AbstractVector.from_.dispatch(precedence=1)
@@ -106,23 +104,23 @@ def from_(cls: type[AbstractVel3D], obj: AbstractVel3D, /) -> AbstractVel3D:
 #####################################################################
 
 
-@AbstractVector.from_.dispatch(precedence=-1)
-def from_(cls: type[AbstractAcc3D], obj: Any, /) -> CartesianAcc3D:
-    """Try to construct a 3D Cartesian velocity from an object.
+# @AbstractVector.from_.dispatch(precedence=-1)
+# def from_(cls: type[AbstractAcc3D], obj: Any, /) -> CartesianAcc3D:
+#     """Try to construct a 3D Cartesian velocity from an object.
 
-    Examples
-    --------
-    >>> import unxt as u
-    >>> import coordinax as cx
+#     Examples
+#     --------
+#     >>> import unxt as u
+#     >>> import coordinax as cx
 
-    >>> x = u.Quantity([1, 2, 3], "km / s2")
-    >>> vec = cx.vecs.AbstractAcc3D.from_(x)
-    >>> print(vec)
-    <CartesianAcc3D (d2_x[km / s2], d2_y[km / s2], d2_z[km / s2])
-        [1 2 3]>
+#     >>> x = u.Quantity([1, 2, 3], "km / s2")
+#     >>> vec = cx.vecs.AbstractAcc3D.from_(x)
+#     >>> print(vec)
+#     <CartesianAcc3D (d2_x[km / s2], d2_y[km / s2], d2_z[km / s2])
+#         [1 2 3]>
 
-    """
-    return obj if isinstance(obj, CartesianAcc3D) else CartesianAcc3D.from_(obj)
+#     """
+#     return obj if isinstance(obj, CartesianAcc3D) else CartesianAcc3D.from_(obj)
 
 
 @AbstractVector.from_.dispatch(precedence=1)
