+++ src/coordinax/_src/vectors/d3/generic.py
@@ -4,14 +4,13 @@ __all__ = [
     "CartesianGeneric3D",
 ]
 
-from dataclasses import fields
 from typing import TypeVar, final
 
 import equinox as eqx
 import jax
 
 import quaxed.numpy as jnp
-from unxt.quantity import AbstractQuantity, Quantity
+from unxt.quantity import Quantity
 
 import coordinax._src.typing as ct
 from coordinax._src.vectors.base import AbstractVector
@@ -72,30 +71,3 @@ class CartesianGeneric3D(AvalMixin, AbstractVector):
 
         """
         return jax.tree.map(jnp.negative, self)
-
-
-# =====================================================
-# Constructors
-
-
-@AbstractVector.from_.dispatch  # type: ignore[misc]
-def from_(
-    cls: type[CartesianGeneric3D],
-    obj: AbstractQuantity,  # TODO: Shaped[AbstractQuantity, "*batch 3"]
-    /,
-) -> CartesianGeneric3D:
-    """Construct a 3D Cartesian position.
-
-    Examples
-    --------
-    >>> import unxt as u
-    >>> import coordinax as cx
-
-    >>> vec = cx.vecs.CartesianGeneric3D.from_(u.Quantity([1, 2, 3], "m"))
-    >>> print(vec)
-    <CartesianGeneric3D (x[m], y[m], z[m])
-        [1 2 3]>
-
-    """
-    comps = {f.name: obj[..., i] for i, f in enumerate(fields(cls))}
-    return cls(**comps)
